% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{addPkg}
\alias{addPkg}
\alias{addPkg,PkgManifest-method}
\alias{addPkg,PkgManifest}
\alias{addPkg,SessionManifest-method}
\alias{addPkg,SessionManifest}
\title{addPkg}
\usage{
addPkg(x, ..., rows = makeManifest(...), versions = data.frame(name =
  manifest_df(rows)$name, version = NA_character_, stringsAsFactors =
  FALSE), replace = FALSE)

\S4method{addPkg}{PkgManifest}(x, ..., rows = makeManifest(...),
  versions = data.frame(name = manifest_df(rows)$name, version =
  NA_character_, stringsAsFactors = FALSE), replace = FALSE)

\S4method{addPkg}{SessionManifest}(x, ..., rows = makeManifest(...),
  versions = data.frame(name = manifest_df(rows)$name, version =
  NA_character_, stringsAsFactors = FALSE), replace = FALSE)
}
\arguments{
\item{x}{A manifest or manifest-associate objec tto add the pkg 2}

\item{\dots}{The information regarding the package to place in the manifest}

\item{rows}{An already-created data.frame to add to the manifest}

\item{versions}{A data.frame of package names and versions, if adding to
a SessionManifest, ignored otherwise}

\item{replace}{logical. If true, the specified package info will replace
any already in the manifest in the case of duplicates. Otherwise, an error
is thrown.}
}
\description{
Add a package to an object associated with a manifest
}
