% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{addReplaceDF}
\alias{addReplaceDF}
\title{Add/replace rows in a data.frame

Combine two dataframes together with rows in one optionally replacing
those in the other when they match on a specified index column}
\usage{
addReplaceDF(df, newdf, replace = TRUE, indexcol = "name")
}
\arguments{
\item{df}{data.frame. The "first" or "old" data.frame.}

\item{newdf}{data.frame. The "new" data frame of rows to add to \code{df}}

\item{replace}{logical. Should replacement happen when rows of \code{df} and
\code{newdf} match based on \code{indexcol}. Defaults to \code{TRUE}. If
\code{FALSE}, an error is thrown in the matching case.}

\item{indexcol}{character. The name of the column to use for matching.
Defaults to \code{"name"} for convenience of internal usage.}
}
\value{
A combined data.frame with only columns found in both data.frames
and one row per unique value of the specified index column across both
datasets.
}

