% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{archive_retries}
\alias{archive_retries}
\alias{archive_retries,SwitchrParam-method}
\alias{archive_retries,SwitchrParam}
\alias{archive_retries<-}
\alias{archive_retries<-,SwitchrParam-method}
\alias{archive_retries<-,SwitchrParam}
\title{archive_retries}
\usage{
archive_retries(x)

\S4method{archive_retries}{SwitchrParam}(x)

archive_retries(x) <- value

\S4method{archive_retries}{SwitchrParam}(x) <- value
}
\arguments{
\item{x}{A SwitchrParam object}

\item{value}{The new number of seconds to wait}
}
\value{
When getting, the number of seconds to wait,
when setting, a new, updated SwitchrParam object.
}
\description{
Get or set the number of times to retry downloading a file from
the CRAN archive

This is intended to stop intermittent install failures
due to failing to retrieve files that *are* in the
archive but are not downloading properly when a larger 
number of packages is being retrieved.
}
