% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeBiocVersions.R
\docType{data}
\name{BiocDevel}
\alias{BiocDevel}
\title{BiocDevel}
\format{\preformatted{Formal class 'RepoSubset' [package "roxygen_devtest"] with 3 slots
  ..@ repos       : Named chr [1:5] "http://bioconductor.org/packages/3.0/bioc" "http://bioconductor.org/packages/3.0/data/annotation" "http://bioconductor.org/packages/3.0/data/experiment" "http://bioconductor.org/packages/3.0/extra" ...
  .. ..- attr(*, "names")= chr [1:5] "BioCsoft" "BioCann" "BioCexp" "BioCextra" ...
  ..@ pkgs        : chr "BiocInstaller"
  ..@ default_name: chr "BioC_3.0"
}}
\usage{
BiocDevel
}
\description{
An object representing the current Bioc devel version. Can be passed to switchTo.
}
\keyword{datasets}

