% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{BiocRelease}
\alias{BiocRelease}
\title{BiocRelease}
\format{\preformatted{Formal class 'RepoSubset' [package "roxygen_devtest"] with 3 slots
  ..@ repos       : Named chr [1:5] "http://bioconductor.org/packages/2.14/bioc" "http://bioconductor.org/packages/2.14/data/annotation" "http://bioconductor.org/packages/2.14/data/experiment" "http://bioconductor.org/packages/2.14/extra" ...
  .. ..- attr(*, "names")= chr [1:5] "BioCsoft" "BioCann" "BioCexp" "BioCextra" ...
  ..@ pkgs        : chr "BiocInstaller"
  ..@ default_name: chr "BioC_2.14"
}}
\usage{
BiocRelease
}
\description{
An object representing the current Bioc release. Can be passed to switchTo.
}
\keyword{datasets}

