% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBiocVersions.R
\name{biocReposForVers}
\alias{biocReposForVers}
\title{biocReposForVers}
\usage{
biocReposForVers(version)
}
\arguments{
\item{version}{The Bioconductor release to generate URLs for.}
}
\description{
Generate the URLs of the repositories associated with
a specific Bioconductor release
}
\note{
This function will only work if some version of Bioconductor (>2.9)
was installed when switchr was installed. It will return NULL otherwise.
}
