% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{branch}
\alias{branch}
\alias{branch,PkgSource-method}
\alias{branch,PkgSource}
\alias{branch<-}
\alias{branch<-,PkgSource-method}
\alias{branch<-,PkgSource}
\title{branch}
\usage{
branch(x)

\S4method{branch}{PkgSource}(x)

branch(x) <- value

\S4method{branch}{PkgSource}(x) <- value
}
\arguments{
\item{x}{A source}

\item{value}{The new branch}
}
\value{
for the getter, the branch associated with the source
object, for the setter, the object updated to use the specified
branch.
}
\description{
Get or set the branch associated with a Package Source
}
