% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{c,SessionManifest-method}
\alias{c,PkgManifest-method}
\alias{c,SessionManifest-method}
\title{Combine package manifests}
\usage{
\S4method{c}{SessionManifest}(x, ..., recursive = FALSE)

\S4method{c}{PkgManifest}(x, ..., recursive = FALSE)
}
\arguments{
\item{x}{A SessionManifest object}

\item{...}{One or more SessionManifest objects}

\item{x}{A PkgManifest object}

\item{...}{One or more PkgManifest objects}

\item{recursive}{logical.  If \code{recursive = TRUE}, the function
    recursively descends through lists (and pairlists) combining all
    their elements into a vector.}
}
\value{
A SessionManifest object

A PkgManifest object
}
\description{
Combine package manifests
}

