% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{c,SessionManifest-method}
\alias{c,SessionManifest-method}
\title{cmethods}
\usage{
\S4method{c}{SessionManifest}(x, ..., recursive = FALSE)
}
\arguments{
\item{...}{objects to be concatenated.}

\item{recursive}{logical.  If \code{recursive = TRUE}, the function
    recursively descends through lists (and pairlists) combining all
    their elements into a vector.}
}
\description{
cmethods
}

