% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/c-methods.R, R/cmethods.R
\docType{methods}
\name{c,SessionManifest-method}
\alias{c,PkgManifest-method}
\alias{c,SessionManifest-method}
\title{cmethods}
\usage{
\S4method{c}{SessionManifest}(x, ..., recursive = FALSE)

\S4method{c}{PkgManifest}(x, ..., recursive = FALSE)

\S4method{c}{SessionManifest}(x, ..., recursive = FALSE)

\S4method{c}{PkgManifest}(x, ..., recursive = FALSE)
}
\description{
cmethods
}

