% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conditionalCode.R
\name{defaultRepos}
\alias{defaultRepos}
\title{defaultRepos}
\usage{
defaultRepos()
}
\value{
A character vector of package repository urls
}
\description{
Get default repositories for use as dependency repos and within
install_packages
}

