% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{dl_method}
\alias{dl_method}
\alias{dl_method,SwitchrParam-method}
\alias{dl_method,SwitchrParam}
\alias{dl_method<-}
\alias{dl_method<-,SwitchrParam-method}
\alias{dl_method<-,SwitchrParam}
\title{dl_method}
\usage{
dl_method(x)

\S4method{dl_method}{SwitchrParam}(x)

dl_method(x) <- value

\S4method{dl_method}{SwitchrParam}(x) <- value
}
\arguments{
\item{x}{A SwitchrParam object}

\item{value}{The new number of seconds to wait}
}
\value{
for the getter, the download method specified in the
\code{SwitchrParam} object, for the setter, the object
updated with the new download method.
}
\description{
Get or set the download method for retreiving files.
}
