% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{fileFromFileURL}
\alias{fileFromFileURL}
\title{Get path from file URL}
\usage{
fileFromFileURL(fileurl)
}
\arguments{
\item{fileurl}{A file url (beginning in file://)}
}
\value{
The system directory path that \code{fileurl} points to
}
\description{
Get path from file URL
}
