% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flushSession}
\alias{flushSession}
\title{Empty current R session}
\usage{
flushSession(dontunload = c("switchr", "digest"))
}
\arguments{
\item{dontunload}{Non-base packages to ignore (not detatch/unload)}
}
\description{
Empty current R session
}

