% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{full_libpaths}
\alias{full_libpaths}
\alias{full_libpaths,SwitchrCtx-method}
\alias{full_libpaths,SwitchrCtx}
\title{full_libpaths}
\usage{
full_libpaths(seed)

\S4method{full_libpaths}{SwitchrCtx}(seed)
}
\arguments{
\item{seed}{a SwitchrCtx}
}
\value{
For the getter, the full set of library paths associated with the
\code{SwitchrCtx} object, for the setter, the object, updated with the
new set of full lib paths.
}
\description{
Accessor for the full library path associate with a SwitchrCtx, including
the R library and (if not excluded) the site library
}
