% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getPkgDir}
\alias{getPkgDir}
\title{Construct pockage directory path}
\usage{
getPkgDir(basepath, name, subdir, scm_type, branch)
}
\arguments{
\item{basepath}{The parent directory for the package directory}

\item{name}{The name of the package}

\item{subdir}{The subdirectory within a package source that
the actual package root directory will reside in.}

\item{scm_type}{Tye type of scm the package sources will be
checked out from}

\item{branch}{The branch from which the package will be retrieved.}
}
\value{
A path
}
\description{
Construct pockage directory path
}
\note{
Unlike \code{\link{findPkgDir}} this does not look for existing
package source directories. It only constructs the path.
}
