% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GithubManifest}
\alias{GithubManifest}
\title{GithubManifest}
\usage{
GithubManifest(..., pkgrepos = as.character(list(...)))
}
\arguments{
\item{pkgrepos}{Github repositories in the form "<user>/<reponame>"}

\item{\dots{}}{Combined to populate \code{pkgrepos}}
}
\description{
Create a package manifest containing only github packages
}
\note{
This is a convenience wrapper for \code{\link{Manifest}}. Non-default
location information (e.g. branches other than master,
subdirectories within the repository) are not currently supported. Use
\code{\link{Manifest}} or edit the package manifest after
creation when those are required.
}

