% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/retrievePkgVersion.R
\docType{methods}
\name{gotoVersCommit}
\alias{gotoVersCommit}
\alias{gotoVersCommit,character,BiocSource}
\alias{gotoVersCommit,character,BiocSource-method}
\alias{gotoVersCommit,character,CRANSource}
\alias{gotoVersCommit,character,CRANSource-method}
\alias{gotoVersCommit,character,GitSource}
\alias{gotoVersCommit,character,GitSource-method}
\alias{gotoVersCommit,character,SVNSource}
\alias{gotoVersCommit,character,SVNSource-method}
\title{gotoVersCommit
This is a low-level function not intended for direct use by the end user.}
\usage{
gotoVersCommit(dir, src, version, param = SwitchrParam())

\S4method{gotoVersCommit}{character,SVNSource}(dir, src, version,
  param = SwitchrParam())

\S4method{gotoVersCommit}{character,CRANSource}(dir, src, version,
  param = SwitchrParam())

\S4method{gotoVersCommit}{character,BiocSource}(dir, src, version,
  param = SwitchrParam())

\S4method{gotoVersCommit}{character,GitSource}(dir, src, version,
  param = SwitchrParam())
}
\arguments{
\item{dir}{Directory}

\item{src}{A PkgSource (or subclass) object}

\item{version}{The exact version to locate}

\item{param}{A SwitchrParam}
}
\description{
gotoVersCommit
This is a low-level function not intended for direct use by the end user.
}

