% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head-methods.R
\docType{methods}
\name{head}
\alias{head}
\alias{head,SessionManifest-method}
\alias{head,SessionManifest}
\alias{head,PkgManifest-method}
\alias{head,PkgManifest}
\alias{tail}
\alias{tail,SessionManifest-method}
\alias{tail,SessionManifest}
\alias{tail,PkgManifest-method}
\alias{tail,PkgManifest}
\title{Head and tail operations on manifests}
\usage{
head(x, ...)

\S4method{head}{SessionManifest}(x, n = 5, ...)

\S4method{head}{PkgManifest}(x, n = 5, ...)

tail(x, ...)

\S4method{tail}{SessionManifest}(x, n = 5, ...)

\S4method{tail}{PkgManifest}(x, n = 5, ...)
}
\arguments{
\item{x}{A manifest object}

\item{...}{unused}

\item{n}{The number of packages to keep}
}
\value{
An object of the same type as \code{x} containing \code{n} packages
}
\description{
Head and tail operations on manifests
}
\details{
In the case of a \code{PkgManifest}, the first or last \code{n}
packages are retained in the manifest, while all others are removed.

In the case of a \code{SessionManifest}, \code{n} specified versions
are retained, while the underlying \code{PkgManifest} is unchanged.
}
