% Generated by roxygen2 (4.0.1): do not edit by hand
\name{installCompEnv}
\alias{installCompEnv}
\title{Install a new set of packages}
\usage{
installCompEnv(repo_base, doi, repo_name = if (!missing(doi)) doi else "",
  pkgs = available.packages(contrib.url(repo))[, "Package"], name,
  exclude.site = TRUE, switchTo = FALSE, deps_repos = c(defGRAN, bioc),
  download.dir = tempdir(), ...)
}
\arguments{
\item{repo_base}{The repo base url. repo_name (which may be "") will be concatenated to it}

\item{doi}{(optional) a DOI to associate with this computing environment (typically the DOI of the work being reproduced)}

\item{repo_name}{The name of the repository (on top of repo_base). Defaults to doi if it is specified, and "" otherwise}

\item{pkgs}{The packages to install from the specified repository. Defaults to all packages in the repo}

\item{name}{The name to associate with the computing environment Defaults to doi if it is non-null, and a hash of the full repository url otehrwise}

\item{exclude.site}{Whether packages installed in the current R_LIBS_SITE location should be counted as installed dependencies}
}
\description{
Install a new set of packages
}
\note{
The default is to install *ALL* packages in the specified repository. This is appropriate for reproducibility-based repositories but NOT for
more standard ones like CRAN or BioConductor.
}

