% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00graceful_inet.R
\name{graceful_inet}
\alias{graceful_inet}
\alias{warning2}
\title{Internal internet harness}
\usage{
graceful_inet(val)

warning2(...)
}
\arguments{
\item{val}{logical. NA means no additional handling, TRUE, means careful handling
but actually attempt the call, FALSE means force immedate failure without evaling
expressions wrapped in inet_handlers() calls}

\item{\dots}{passed to message or base::warning}
}
\description{
This function should never be called by code outside of tests/vignettes
in this package or packages that depend on it.
}
