% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/manifestFromLib.R
\name{libManifest}
\alias{libManifest}
\title{libManifest
Create a Session- or PkgManifest for the contents of a switchr
library.}
\usage{
libManifest(lib = currentCompEnv(), record_versions = TRUE,
  known_manifest = Manifest(dep_repos = repos), repos = defaultRepos(), ...)
}
\arguments{
\item{lib}{A SwitchrCtx object. Defaults to the currently active switchr
library.}

\item{record_versions}{Should the exact versions of installed packages be
recorded in the manifest (TRUE)}

\item{known_manifest}{An existing manifest, used when imputing
location information for packages not
installed via \code{\link{install_packages}}}

\item{A}{vector of traditional package repositories. Used when imputing
location information for packages not installed via
\code{\link{install_packages}}}

\item{\dots{}}{currently unused}
}
\description{
libManifest
Create a Session- or PkgManifest for the contents of a switchr
library.
}

