% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accessors.R
\name{library_paths}
\alias{library_paths}
\title{library_paths}
\usage{
library_paths(seed)
}
\arguments{
\item{seed}{An SwitchrCtx}
}
\description{
Accessor for which directories an SwitchrCtx is associated with.
}

