% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.libPaths2}
\alias{.libPaths2}
\title{.libpaths2}
\usage{
.libPaths2(fulllp, exclude.site = TRUE)
}
\arguments{
\item{fulllp}{The libpath to use, as in .libPaths}

\item{exclude.site}{logical. Should the site library be suppressed.
Defaults to TRUE}
}
\description{
A version of .libPaths which allows for excluding the site library
}
\details{
Behaves exactly as the .libPaths function does, with the exception
of optionally excluding the site library
}
