% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGRAN.R
\name{loadGRAN}
\alias{loadGRAN}
\title{Load a GRAN repo package}
\usage{
loadGRAN(nm = "current")
}
\arguments{
\item{nm}{The name of the repository for which to load the package. Defaults
\code{"current"}}
}
\value{
NULL. Called for the side-effect of loading the specified package
}
\description{
Load a GRAN repo package
}
\details{
This function is a convenience to load the package GRAN<nm>, which
will provide the contained GRAN repository as default repository within
the switchr framework.
}
