% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadManifest}
\alias{loadManifest}
\title{loadManifest}
\usage{
loadManifest(fil)
}
\arguments{
\item{fil}{The path or URL to the file or a gist containing it}
}
\value{
A PkgManifest or SessionManifest object
}
\description{
Load a package or session manifest from a file (local or URL)
}

