% Generated by roxygen2 (4.0.1): do not edit by hand
\name{locatePkgVersion}
\alias{locatePkgVersion}
\title{locatePkgVersion}
\usage{
locatePkgVersion(name, version, pkg_manifest, param = SwitchrParam(),
  dir = if (is.null(repo)) tempdir() else notrack(repo), repo = NULL)
}
\arguments{
\item{name}{package name}

\item{version}{package version string}

\item{repo}{(optional) GRANRepository object to search}

\item{dir}{directory to download package into}
}
\value{
The full path to the downloaded file , or NULL if unable to
locate the package
}
\description{
Locate and download/build the exact version of a single package.
}
\author{
Gabriel Becker
}

