% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{logfun}
\alias{logfun}
\alias{logfun,SwitchrParam}
\alias{logfun,SwitchrParam-method}
\alias{logfun<-}
\alias{logfun<-,SwitchrParam}
\alias{logfun<-,SwitchrParam-method}
\title{Get or set the logging function in an object associated with a SwitchrParam}
\usage{
logfun(x)

\S4method{logfun}{SwitchrParam}(x)

logfun(x) <- value

\S4method{logfun}{SwitchrParam}(x) <- value
}
\arguments{
\item{x}{An object with a SwitchrParam}

\item{value}{The new logging function}
}
\description{
Get or set the logging function in an object associated with a SwitchrParam
}

