% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/retrievePkgVersion.R
\name{makeBiocSVNURL}
\alias{makeBiocSVNURL}
\title{Make a Bioconductor SVN url for a package}
\usage{
makeBiocSVNURL(name, biocVers = "trunk")
}
\arguments{
\item{name}{The name of the package}

\item{biocVers}{The version (release) of bioconductor, or trunk (the default) for
Bioc devel.}
}
\description{
Make a Bioconductor SVN url for a package
}

