% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/makePkgSourceDir-methods.R
\docType{methods}
\name{makePkgDir}
\alias{makePkgDir}
\alias{makePkgDir,ANY,SVNSource-method}
\alias{makePkgDir,ANY,SVNSource}
\alias{makePkgDir,ANY,GithubSource-method}
\alias{makePkgDir,ANY,GithubSource}
\alias{makePkgDir,ANY,GitSource-method}
\alias{makePkgDir,ANY,GitSource}
\alias{makePkgDir,ANY,ANY-method}
\alias{makePkgDir,ANY,ANY}
\alias{makePkgDir,ANY,CRANSource-method}
\alias{makePkgDir,ANY,CRANSource}
\alias{makePkgDir,ANY,BiocSource-method}
\alias{makePkgDir,ANY,BiocSource}
\alias{makePkgDir,ANY,TarballSource-method}
\alias{makePkgDir,ANY,TarballSource}
\alias{makePkgDir,ANY,LocalSource-method}
\alias{makePkgDir,ANY,LocalSource}
\title{makePkgDir}
\usage{
makePkgDir(name, source, path, latest_only, param = SwitchrParam(),
  forceRefresh = FALSE)

\S4method{makePkgDir}{ANY,SVNSource}(name, source, path,
  latest_only = FALSE, param = SwitchrParam(), forceRefresh = FALSE)

\S4method{makePkgDir}{ANY,GithubSource}(name, source, path,
  latest_only = FALSE, param = SwitchrParam(), forceRefresh = FALSE)

\S4method{makePkgDir}{ANY,GitSource}(name, source, path,
  latest_only = FALSE, param = SwitchrParam(), forceRefresh = FALSE)

\S4method{makePkgDir}{ANY,ANY}(name, source, path, latest_only,
  param = SwitchrParam(), forceRefresh = FALSE)

\S4method{makePkgDir}{ANY,CRANSource}(name, source, path, latest_only,
  param = SwitchrParam(), forceRefresh = FALSE)

\S4method{makePkgDir}{ANY,BiocSource}(name, source, path, latest_only,
  param = SwitchrParam(), forceRefresh = FALSE)

\S4method{makePkgDir}{ANY,TarballSource}(name, source, path, latest_only,
  param = SwitchrParam(), forceRefresh = FALSE)

\S4method{makePkgDir}{ANY,LocalSource}(name, source, path, latest_only,
  param = SwitchrParam(), forceRefresh = FALSE)
}
\arguments{
\item{name}{The package}

\item{source}{A PkgSource}

\item{path}{The path to place the directory}

\item{latest_only}{Should a fastpath for downloading the latest commit
in a SCM package without a formal checkout be used?}

\item{param}{A SwitchrParam}

\item{forceRefresh}{Should an existing instance of the package source be
deleted/refreshed}
}
\description{
This is an internal function not intended to be called directly by end users
}
\details{
Create a directory and populate it with package source code from the
specified source
}
