% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeSeedMan}
\alias{makeSeedMan}
\title{makeSeedMan}
\usage{
makeSeedMan(x, known_manifest = PkgManifest(), ...)
}
\arguments{
\item{x}{The object to generate a seeding manifest from}

\item{known_manifest}{A manifest containing known locations of package sources.
makeSeedMan will attempt to determine locations of packages listed in x using both
known_manifest and official repositories.}
}
\description{
makeSeedMan
}

