% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{makeSource}
\alias{makeSource}
\title{Create a PkgSource object for a package}
\usage{
makeSource(url, type, user, password, scm_auth = list(),
  prefer_svn = FALSE, ...)
}
\arguments{
\item{url}{The url of the package sources}

\item{type}{The source type.}

\item{user}{A function which, when called, returns the username to use when
when checking the soources out}

\item{password}{A function which returns the password to use when checking
out the sources}

\item{scm_auth}{A list of username-password pairs, named with regular
expressions to match against url when constructing the
defaults for \code{user} and \code{password}}

\item{prefer_svn}{Currently unused.}

\item{\dots}{Passed directly to constructors for PkgSource superclasses}
}
\description{
Create a PkgSource object for a package
}
