% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/manifest.R
\name{Manifest}
\alias{Manifest}
\title{Manifest}
\usage{
Manifest(..., dep_repos = c(biocinstallRepos()))
}
\arguments{
\item{dep_repos}{The dependency repos for the package.}

\item{\dots{}}{Vectors containing package information. Passed to \code{\link{ManifestRow}}}
}
\description{
Create a package manifest
}

