% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accessors.R
\name{manifest_df}
\alias{manifest_df}
\alias{manifest_df<-}
\title{manifest_df}
\usage{
manifest_df(x, ...)

manifest_df(x) <- value
}
\arguments{
\item{x}{The object}

\item{value}{A data.frame of package manifest information.
See \code{\link{ManifestRow}}}

\item{\dots{}}{unused.}
}
\description{
Get or set the package location manifest (data.frame) associated with an
object
}

