% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ManifestRow}
\alias{ManifestRow}
\title{ManifestRow}
\usage{
ManifestRow(name = NA, url = NA, type = NA, branch = NA, subdir = ".",
  extra = NA)
}
\arguments{
\item{name}{name of the package}

\item{url}{location of the package sources}

\item{type}{type of location (svn, git, local, etc)}

\item{branch}{name of the branch to use to build the package}

\item{subdir}{subdirectory to use to build the package}

\item{extra}{currently ignored. extra commands for building or
installing the package}
}
\value{
A valid Package manifest data.frame
}
\description{
Create one or more rows of manifest
}

