% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utilities.R
\name{normalizePath2}
\alias{normalizePath2}
\title{normalizePath2}
\usage{
normalizePath2(path, follow.symlinks = FALSE, winslash = "\\\\",
  mustWork = NA)
}
\arguments{
\item{path}{The path to normalize}

\item{follow.symlinks}{Should symlinks (other than . and ..)
be resolved to their physical locations? (FALSE)}

\item{winslash}{The value of winslash to be passed down to normalizePath
on windows systems}

\item{mustWork}{logical. Behavior to perform in the case of a non-existent path.
FALSE does nothing, TRUE throws an error, NA throws a warning.}
}
\value{
The normalized path.
}
\description{
Attempt to normalize a relative path to an absolute one. Optionally without
resolving symlinks on non-Windows systems
}

