% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accessors.R
\name{notrack}
\alias{notrack}
\title{Notrack directory}
\usage{
notrack(repo)
}
\arguments{
\item{repo}{The object.}
}
\value{
the path where retrieved package versions should be. If \code{repo}
is NULL, a notrack directory is constructed within a temp directory.
}
\description{
This function is not intended to be called directly by the user.
}

