% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{nrow}
\alias{nrow}
\title{Number of rows}
\usage{
nrow(x)
}
\arguments{
\item{x}{A tabular data structure.}
}
\value{
The number of rows in the structure
}
\description{
Number of rows
}

