% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{packages}
\alias{packages}
\alias{packages,SwitchrCtx}
\alias{packages,SwitchrCtx-method}
\title{List the packages installed in a switchr context (library)}
\usage{
packages(seed)

\S4method{packages}{SwitchrCtx}(seed)
}
\arguments{
\item{seed}{A switchr context}
}
\description{
List the packages installed in a switchr context (library)
}

