% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parseSessionInfoString}
\alias{parseSessionInfoString}
\title{Parse text output from printing SessionInfo objects}
\usage{
parseSessionInfoString(string)
}
\arguments{
\item{string}{The text output from sessionInfo()}
}
\description{
Parse text output from printing SessionInfo objects
}

