% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{PkgManifest}
\alias{PkgManifest}
\title{PkgManifest}
\usage{
PkgManifest(manifest = ManifestRow(...), dep_repos = c(biocinstallRepos(),
  defaultGRAN()), ...)
}
\arguments{
\item{manifest}{The manifest (data.frame) of packages and their locations}

\item{dep_repos}{A list of traditional pkg repositories which can contain dependencies
for the packages listed in \code{manifest}.}
}
\description{
Construct a PkgManifest, which can be installed from using \code{\link{install_packages}}
}
\details{
If a package is found in both the manifest dataf.frame and the dependency
repositories, the version in the manifest will always take precidence within the
switchr framework.
}

