% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{pkgname}
\alias{pkgname}
\alias{pkgname,PkgSource-method}
\alias{pkgname,PkgSource}
\alias{pkgname<-}
\alias{pkgname<-,PkgSource-method}
\alias{pkgname<-,PkgSource}
\title{pkgname}
\usage{
pkgname(x)

\S4method{pkgname}{PkgSource}(x)

pkgname(x) <- value

\S4method{pkgname}{PkgSource}(x) <- value
}
\arguments{
\item{x}{A source}

\item{value}{The new pkgname}
}
\value{
for the getter, the package name as a string, for the setter,
an updated \code{PkgSource} (or subclass) object with the new
package name.
}
\description{
Get or set the package name associated with a Package Source
}
