% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{PkgSource-class}
\alias{PkgSource-class}
\title{PkgSource
An object representing the source location of a package. This is a virtual
used exlusively through its subclasses, which are used to differentiate the
different types of package source locations.}
\description{
PkgSource
An object representing the source location of a package. This is a virtual
used exlusively through its subclasses, which are used to differentiate the
different types of package source locations.
}

