% Generated by roxygen2 (4.0.1): do not edit by hand
\name{publishManifest}
\alias{publishManifest}
\title{publishManifest}
\usage{
publishManifest(manifest, dest = "./pkg_manifest.rman", ...)
}
\arguments{
\item{manifest}{The object to save as a serialized package or session
manfiest. A session manifest will be generated by libManifest as necessary.}

\item{dest}{Name (including path) of the file to save}

\item{...}{Unused}
}
\value{
The name of the file written
}
\description{
Pub a package or session manifest to file.
}

