% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Rcmd}
\alias{Rcmd}
\title{R executable}
\usage{
Rcmd(cmd = c("build", "check", "INSTALL", ""), options)
}
\arguments{
\item{cmd}{the R CMD to run. "build", "check", "INSTALL", or "" (for none)}

\item{options}{The options to pass to the command}
}
\description{
R executable
}

