% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readManifest}
\alias{readManifest}
\title{readManifest}
\usage{
readManifest(uri, local = !url.exists(uri), archive = FALSE)
}
\arguments{
\item{uri}{The location of the manifest directory (path or URL)}

\item{local}{Whether the manifest is a local directory or a URL}

\item{archive}{Not currently supported}
}
\value{
A PackageManifest object, or a SessionManifest object if the
manifest directory contains a pkg_versions.dat file.
}
\description{
Read a package or session manifest from a remote or local directory
}

