% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/removeLibrary.R
\name{removeLib}
\alias{removeLib}
\title{removeLib
Remove a switchr library and update the manifest of existing libraries}
\usage{
removeLib(name = NULL, repos = NULL, compEnv = NULL, fromStack = FALSE)
}
\value{
NULL, called for its side-effect of removing/destroying a switchr
library
}
\description{
removeLib
Remove a switchr library and update the manifest of existing libraries
}

