% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{RepoSubset-class}
\alias{RepoSubset-class}
\title{RepoSubset}
\description{
An object that represents a subset of packages available in a repo. When switched to, switchr will default to only installing the
specified packages, rather than all packages in the repository.
}

