% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crandbtmp.R
\name{rVersionManifest}
\alias{rVersionManifest}
\title{rVersionManifest
Create a Pkg manifest which points to tarballs representing the
cohort of packages associated with a particular release of R}
\usage{
rVersionManifest(vers, curr_avail = available.packages())
}
\arguments{
\item{vers}{The version of R to create a manifest for}

\item{curr_avail}{The output from available.packages(). Used to identify
whether the necessary version is in the CRAN archive or normal repository}
}
\value{
A SessionManifest object
}
\description{
rVersionManifest
Create a Pkg manifest which points to tarballs representing the
cohort of packages associated with a particular release of R
}
\examples{
\dontrun{
man = rVersionManifest("3.1.1")
man
}
}
\author{
Gabriel Becker
}
\references{
"Gabor Csardi" (2014). crandb: Query the unofficial CRAN metadata
 database. R package version 1.0.0. https://github.com/metacran/crandb
}

