% Generated by roxygen2 (4.0.1): do not edit by hand
\name{saveManifest}
\alias{saveManifest}
\title{saveManifest}
\usage{
saveManifest(manifest, fname = "./pkg_manifest.rman", ...)
}
\arguments{
\item{manifest}{The object to save as a serialized package or session
manfiest. A session manifest will be generated by libManifest as necessary.}

\item{fname}{Name (including path) of the file to save}

\item{...}{Unused}
}
\value{
The name of the file written
}
\description{
Save a package or session manifest to file.
}

