% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{sh_init_script}
\alias{sh_init_script}
\alias{sh_init_script,SwitchrParam}
\alias{sh_init_script,SwitchrParam-method}
\alias{sh_init_script<-}
\alias{sh_init_script<-,SwitchrParam,ANY}
\alias{sh_init_script<-,SwitchrParam-method}
\title{Set or Retrieve the shell initialization script for an object}
\usage{
sh_init_script(x)

\S4method{sh_init_script}{SwitchrParam}(x)

sh_init_script(x) <- value

\S4method{sh_init_script}{SwitchrParam}(x) <- value
}
\arguments{
\item{x}{An object associated with a SwitchrParam object}

\item{value}{The new value.}
}
\description{
Set or Retrieve the shell initialization script for an object
}

