% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{shell_timing}
\alias{shell_timing}
\alias{shell_timing,SwitchrParam-method}
\alias{shell_timing,SwitchrParam}
\alias{shell_timing<-}
\alias{shell_timing<-,SwitchrParam-method}
\alias{shell_timing<-,SwitchrParam}
\title{Get or set the number of seconds to wait between successive shell commands}
\usage{
shell_timing(x)

\S4method{shell_timing}{SwitchrParam}(x)

shell_timing(x) <- value

\S4method{shell_timing}{SwitchrParam}(x) <- value
}
\arguments{
\item{x}{A SwitchrParam object}

\item{value}{The new number of seconds to wait}
}
\value{
When getting, the number of seconds to wait,
when setting, a new, updated SwitchrParam object.
}
\description{
This is intended to stop intermittent install failures
due to network drive latency interacting with git commands
}
