% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{subdir}
\alias{subdir}
\alias{subdir,PkgSource-method}
\alias{subdir,PkgSource}
\alias{subdir<-}
\alias{subdir<-,PkgSource-method}
\alias{subdir<-,PkgSource}
\title{subdir}
\usage{
subdir(x)

\S4method{subdir}{PkgSource}(x)

subdir(x) <- value

\S4method{subdir}{PkgSource}(x) <- value
}
\arguments{
\item{x}{An object associated with a subdirectory, typically a PkgSource}

\item{value}{The new subdirectory to associate with the object}
}
\value{
For the getter, the subdirectory within the overall source to
find the actual package source code, for the setter, an updated
\code{PkgSource} object with the new subdirectory set.
}
\description{
accessor for subdirectory.
}
