% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{switchBack}
\alias{switchBack}
\title{switchBack}
\usage{
switchBack()
}
\value{
silently, the name of the switchr context now active.
}
\description{
A convenience function to switch back to the previously used computing
environment.
}
