% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{switchDeps}
\alias{switchDeps}
\title{switchrDeps
The base packages, as well as switchr and its dependencies.}
\format{\preformatted{ Named chr [1:19] "base" "compiler" "datasets" "graphics" ...
 - attr(*, "names")= chr [1:19] "base" "compiler" "datasets" "graphics" ...
}}
\usage{
switchDeps
}
\description{
switchrDeps
The base packages, as well as switchr and its dependencies.
}
\keyword{datasets}

