% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{SwitchrCtx}
\alias{SwitchrCtx}
\alias{SwitchrCtx-class}
\title{SwitchrCtx}
\usage{
SwitchrCtx(name, libpaths, exclude.site = TRUE, seed = NULL)
}
\arguments{
\item{name}{The name to associate with the context}

\item{libpaths}{The directories where the installed packages are located}

\item{exclude.site}{Should the current site library be included in the
context when it is switched to (TRUE) '}

\item{seed}{An object representing the list of packages the switchr context
was seeded with.}
}
\description{
A constructor for class SwitchrCtx, represenging a switchr installed-package library.
}

