% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCtx.R
\name{switchrManifest}
\alias{switchrManifest}
\title{switchrManifest}
\usage{
switchrManifest()
}
\value{
A data.frame with information about the located switchr libraries
}
\description{
Generate a manifest of all currently available (existing) swtichr libraries.
}
\note{
This function reads cached metadata from the current switchr base
directory (~/.switchr by default). This cache is updated whenever
the switchr framework is used to create or destroy a switchr library,
but will not be updated if one is added or removed manually. In
such cases \code{\link{updateManifest}} must be called first
}
