% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flushSession.R
\name{switchrNoUnload}
\alias{switchrNoUnload}
\title{Skip unloading of packages in session}
\usage{
switchrNoUnload(value)
}
\arguments{
\item{value}{A logical value, or missing to return the current option}
}
\value{
A logical indicating whether or not calling \code{flushSession} will skipped during the library switching process.
}
\description{
Set whether or not ANY packages are unloaded when switching libraries.
}
\details{
This should be set to TRUE  when using switchr in the context of dynamic documents such as .Rnw and .Rmd files.
}
