% Generated by roxygen2 (4.0.1): do not edit by hand
\name{system_w_init}
\alias{system_w_init}
\title{system_w_init}
\usage{
system_w_init(cmd, init = character(), ..., param = SwitchrParam())
}
\arguments{
\item{cmd}{The text of the command. Must be length 1.}

\item{init}{(optional) a character value indicating the
location of an initialization shell script.}

\item{\dots{}}{additional parameters passed directly to \code{\link{system}}.}

\item{param}{A SwitchrParam object. The shell initialization
script associated with this object is used when \code{init} is
not specified (length 0).}
}
\value{
Depends, see \code{\link{system}} for details.
}
\description{
Run a system command with an optional intialization script (e.g. a .bashrc
sourced first).
}

