% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_PACKAGES2.R
\name{update_PACKAGES}
\alias{update_PACKAGES}
\title{update existing package repository}
\usage{
update_PACKAGES(dir = ".", fields = NULL, type = c("source", "mac.binary",
  "win.binary"), verbose = FALSE, unpacked = FALSE, subdirs = FALSE,
  latestOnly = TRUE, addFiles = FALSE, rds_compress = "xz",
  strict = TRUE, md5_miss_ok = FALSE)
}
\arguments{
\item{dir}{See \code{write_PACKAGES}}

\item{fields}{See \code{write_PACKAGES}}

\item{type}{See \code{write_PACKAGES}}

\item{verbose}{See \code{write_PACKAGES}}

\item{unpacked}{See \code{write_PACKAGES}}

\item{subdirs}{See \code{write_PACKAGES}}

\item{latestOnly}{See \code{write_PACKAGES}}

\item{addFiles}{See \code{write_PACKAGES}}

\item{rds_compress}{See \code{write_PACKAGES}}

\item{strict}{logical. Should all package tarballs without existing matching
entries (by MD5) be processed. if \code{FALSE} package name and version
are extracted from the file name and assumed to be true (This saves timef
for \code{latestOnly = TRUE}.}

\item{md5_miss_ok}{should processing continue if a package version has
an existing PACKAGES entry and a file in \code{dir} with mismatching MD5
sums. If \code{FALSE} (default) an error is thrown, otherwise a warning.}
}
\description{
Update an existing repository by reading the PACKAGES
file and only processing built package tarballs which do not
match existing entries.

\code{update_PACKAGES} can be much faster than \code{write_PACKAGES} for
(relatively) small changes to large repositories.
}
