% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateManifest}
\alias{updateManifest}
\title{updateManifest}
\usage{
updateManifest()
}
\value{
NULL, used for it's side-effect of updating the switchr library
metadata cache.
}
\description{
Update the cached information regarding available switchr libraries.
}

