% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBiocVersions.R
\docType{data}
\name{BiocDevel}
\alias{BiocDevel}
\title{BiocDevel}
\format{An object of class \code{RepoSubset} of length 1.}
\usage{
BiocDevel
}
\description{
An object representing the current Bioc devel version. Can be passed to switchTo.
}
\keyword{datasets}
