% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeBiocVersions.R
\docType{data}
\name{BiocRelease}
\alias{BiocRelease}
\title{BiocRelease}
\format{\preformatted{Formal class 'RepoSubset' [package "roxygen_devtest"] with 3 slots
  ..@ repos       : Named chr [1:5] "http://bioconductor.org/packages/3.1/bioc" "http://bioconductor.org/packages/3.1/data/annotation" "http://bioconductor.org/packages/3.1/data/experiment" "http://bioconductor.org/packages/3.1/extra" ...
  .. ..- attr(*, "names")= chr [1:5] "BioCsoft" "BioCann" "BioCexp" "BioCextra" ...
  ..@ pkgs        : chr "BiocInstaller"
  ..@ default_name: chr "BioC_3.1"
}}
\usage{
BiocRelease
}
\description{
An object representing the current Bioc release. Can be passed to switchTo.
}
\keyword{datasets}

