% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBiocVersions.R
\docType{data}
\name{BiocRelease}
\alias{BiocRelease}
\title{BiocRelease}
\format{An object of class \code{RepoSubset} of length 1.}
\usage{
BiocRelease
}
\description{
An object representing the current Bioc release. Can be passed to switchTo.
}
\keyword{datasets}
