% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeBiocVersions.R
\name{BiocVers}
\alias{BiocVers}
\title{BiocVers
A constructor for creating a RepoSubset object for a
specified release of Bioconductor, which includes only the
BiocInstaller package.}
\usage{
BiocVers(version = biocReleaseV, name = paste("BioC", version, sep = "_"),
  repos = biocReposForVers(version))
}
\description{
BiocVers
A constructor for creating a RepoSubset object for a
specified release of Bioconductor, which includes only the
BiocInstaller package.
}

