% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBiocVersions.R
\name{BiocVers}
\alias{BiocVers}
\title{BiocVers}
\usage{
BiocVers(version = getBiocReleaseVr(), name = paste("BioC", version,
  sep = "_"), repos = biocReposForVers(version))
}
\arguments{
\item{version}{The version of Bioconductor}

\item{name}{The default name for switchr libraries created with this object}

\item{repos}{The urls of the Bioconductor repositories. these will be
modified automatically to match the specified version}
}
\description{
A constructor for creating a RepoSubset object for a
specified release of Bioconductor, which includes only the
BiocInstaller package.
}
