% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{checkIsPkgDir}
\alias{checkIsPkgDir}
\title{Check if a directory contains package sources}
\usage{
checkIsPkgDir(dir)
}
\arguments{
\item{dir}{The directory}
}
\description{
Check if a directory contains package sources
}
