% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmethods.R
\docType{methods}
\name{c,SessionManifest-method}
\alias{c,PkgManifest}
\alias{c,PkgManifest-method}
\alias{c,SessionManifest-method}
\title{cmethods
Combine 2 or more manifests of the same type (PkgManifest or SessionManifest)}
\usage{
\S4method{c}{SessionManifest}(x, ..., recursive = FALSE)

\S4method{c}{PkgManifest}(x, ..., recursive = FALSE)
}
\arguments{
\item{x}{An object}

\item{recursive}{Unused}

\item{\dots}{more objects}
}
\description{
cmethods
Combine 2 or more manifests of the same type (PkgManifest or SessionManifest)
}

