% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmethods.R
\docType{methods}
\name{c,SessionManifest-method}
\alias{c,SessionManifest-method}
\alias{c,PkgManifest-method}
\title{cmethods}
\usage{
\S4method{c}{SessionManifest}(x, ..., recursive = FALSE)

\S4method{c}{PkgManifest}(x, ..., recursive = FALSE)
}
\arguments{
\item{x}{An object (indicates the type of all objects to be combined)}

\item{\dots}{more objects}

\item{recursive}{Unused}
}
\description{
Combine 2 or more manifests of the same type (PkgManifest or SessionManifest)
}
