% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{currentCompEnv}
\alias{currentCompEnv}
\title{currentCompEnv}
\usage{
currentCompEnv()
}
\description{
Display the computing environment currently in use. If switchTo has not been
called, a new SwitchrCtx object describing the current environment is
created.
}
