% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dep_repos}
\alias{dep_repos}
\alias{dep_repos<-}
\title{dep_repos}
\usage{
dep_repos(x)

dep_repos(x) <- value
}
\value{
Character vector with existing repository urls
}
\description{
Get or set repositories to be used to fullfill dependencies beyond packages
within the manifest


}

