% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{errorOrNonZero}
\alias{errorOrNonZero}
\title{Identify error states from R or external programs}
\usage{
errorOrNonZero(out)
}
\arguments{
\item{out}{An R object representing output}
}
\value{
TRUE if out is an error object, or has an attribute called "status" which is > 0
}
\description{
Identify error states from R or external programs
}
