% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flushSession.R
\name{flushSession}
\alias{flushSession}
\title{flushSession}
\usage{
flushSession(dontunload = switchrDontUnload())
}
\arguments{
\item{dontunload}{Non-base packages to ignore (not detatch/unload)}
}
\value{
NULL, called for its side-effect of unloading packages
}
\description{
Unload currently loaded packages from the current R session
}
\details{
Attached packages are detached (and unloaded) first. After this is
done, loaded packages, such as those imported by (previously) attached
packages, are unloaded.

Finally, after all packages have been unloaded, native libraries
loaded by those packages are unloaded (on systems where this is supported).
}
\note{
Failing to include switchr, any of its dependencies, or any base
packages (available as a vector in the \code{\link{switchDeps}} object)
in \code{dontunload} will result in undefined, likely erroneous behavior.
}
