% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{full_libpaths}
\alias{full_libpaths}
\alias{full_libpaths,SwitchrCtx-method}
\alias{full_libpaths,SwitchrCtx}
\title{full_libpaths}
\usage{
full_libpaths(seed)

\S4method{full_libpaths}{SwitchrCtx}(seed)
}
\arguments{
\item{seed}{a SwitchrCtx}
}
\description{
Accessor for the full library path associate with a SwitchrCtx, including
the R library and (if not excluded) the site library
}
