% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{GithubManifest}
\alias{GithubManifest}
\title{GithubManifest}
\usage{
GithubManifest(..., pkgrepos)
}
\arguments{
\item{...}{Combined to populate \code{pkgrepos}}

\item{pkgrepos}{Github repositories in the form "<user>/<reponame>"}
}
\description{
Create a package manifest containing only github packages
}
\details{
Any names of the pkgrepos vector are assumed to be
pkg names for the manifest. For unnamed elements, the pkg
name is assumed to be the repository name.
}
\note{
This is a convenience wrapper for \code{\link{makeManifest}}.
It uses the \code{username/repo[/subdir][@ref]} shorthand for specifying
package locations in github repositories introduced by Wickham's
devtools. Unlike devtools, username is not optional, and  only branch
names are currently supported in the \code{@ref}
}
\examples{
ghman = GithubManifest("gmbecker/switchr", "hadley/devtools")
ghman


}
