% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{library_paths}
\alias{library_paths}
\alias{library_paths,SwitchrCtx-method}
\alias{library_paths,SwitchrCtx}
\title{library_paths}
\usage{
library_paths(seed)

\S4method{library_paths}{SwitchrCtx}(seed)
}
\arguments{
\item{seed}{An SwitchrCtx}
}
\description{
Accessor for which directories an SwitchrCtx is associated with.
}
