% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadManifest.R
\name{loadManifest}
\alias{loadManifest}
\title{loadManifest}
\usage{
loadManifest(fil)
}
\arguments{
\item{fil}{The path or URL to the file or a gist containing it}
}
\value{
A PkgManifest or SessionManifest object
}
\description{
Load a package or session manifest from a file (local or URL)
}
