% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrievePkgVersion.R
\name{locatePkgVersion}
\alias{locatePkgVersion}
\title{locatePkgVersion}
\usage{
locatePkgVersion(name, version, pkg_manifest, param = SwitchrParam(),
  dir = notrack(repo), repo = NULL)
}
\arguments{
\item{name}{package name}

\item{version}{package version string}

\item{pkg_manifest}{A manifest containing locations to search for the
package}

\item{param}{A SwitchrParam object}

\item{dir}{directory to download package into}

\item{repo}{(optional) GRANRepository object to search}
}
\value{
The full path to the downloaded file , or NULL if unable to
locate the package
}
\description{
Locate and download/build the exact version of a single package.
}
\note{
Locating and attempting to install a non-current version of a single
will not work in general, due to dependency issues. In most cases a
Just-in-Time repository should be created and used instead, e.g. via
\code{\link{install_packages}}

This function is called internally during the construction of Just-in-Time
repositories and during the installation of specific package versions.
}
\author{
Gabriel Becker
}
