% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{location}
\alias{location}
\alias{location,PkgSource-method}
\title{location generic
Retreive the directory associated with an object}
\usage{
location(repo)

\S4method{location}{PkgSource}(repo)
}
\arguments{
\item{repo}{An object associated with a path}
}
\value{
a character containing the associated path
}
\description{
location generic
Retreive the directory associated with an object

Retreive the local directory associated with a PkgSource object for a package in a GRAN manifest
}
\author{
Gabriel Becker
}

