% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{location}
\alias{location}
\alias{location,PkgSource-method}
\title{location}
\usage{
location(repo)

\S4method{location}{PkgSource}(repo)
}
\arguments{
\item{repo}{An object associated with a path}
}
\value{
a character containing the associated path
}
\description{
Retreive the directory associated with an object
}
\author{
Gabriel Becker
}
