% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/retrievePkgVersion.R
\name{makeBiocSVNURL}
\alias{makeBiocSVNURL}
\title{Make a Bioconductor SVN url for a package}
\usage{
makeBiocSVNURL(name, biocVers = "devel")
}
\arguments{
\item{name}{A vector of bioconductor package names The name of the package}

\item{biocVers}{The version (release) of bioconductor, or \code{'trunk'} (the default) for
Bioc devel.}
}
\value{
A vector of urls for the specified packages within the Bioconductor SVN repository
}
\description{
Make a Bioconductor SVN url for a package
}

