% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/manifest.R
\name{makeManifest}
\alias{makeManifest}
\title{Manifest constructor
Create a package manifest}
\usage{
makeManifest(..., dep_repos = defaultRepos())
}
\arguments{
\item{dep_repos}{The dependency repos for the package.}

\item{\dots}{Vectors containing package information. Passed to \code{\link{ManifestRow}}}
}
\description{
Manifest constructor
Create a package manifest
}

