% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{makeManifest}
\alias{makeManifest}
\title{Manifest constructor}
\usage{
makeManifest(..., dep_repos = defaultRepos())
}
\arguments{
\item{...}{Vectors containing package information. Passed to \code{\link{ManifestRow}}}

\item{dep_repos}{The dependency repos for the package.}
}
\description{
Create a package manifest
}
