% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manifestFromLib.R
\docType{methods}
\name{makeSeedMan}
\alias{makeSeedMan}
\alias{makeSeedMan,parsedSessionInfo}
\alias{makeSeedMan,parsedSessionInfo-method}
\title{makeSeedMan}
\usage{
makeSeedMan(x, known_manifest = PkgManifest(), ...)

\S4method{makeSeedMan}{parsedSessionInfo}(x, known_manifest = PkgManifest(),
  ...)
}
\arguments{
\item{x}{The object to generate a seeding manifest from}

\item{known_manifest}{A manifest containing known locations of package sources.
makeSeedMan will attempt to determine locations of packages listed in x using both
known_manifest and official repositories.}

\item{...}{Currently unused.}
}
\description{
makeSeedMan
}

