% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifestFromLib.R
\docType{methods}
\name{makeSeedMan}
\alias{makeSeedMan}
\alias{makeSeedMan,missing-method}
\alias{makeSeedMan,missing}
\alias{makeSeedMan,sessionInfo-method}
\alias{makeSeedMan,sessionInfo}
\alias{makeSeedMan,parsedSessionInfo-method}
\alias{makeSeedMan,parsedSessionInfo}
\alias{makeSeedMan,data.frame-method}
\alias{makeSeedMan,data.frame}
\title{makeSeedMan}
\usage{
makeSeedMan(x, known_manifest = PkgManifest(), ...)

\S4method{makeSeedMan}{missing}(x, known_manifest = PkgManifest(), ...)

\S4method{makeSeedMan}{sessionInfo}(x, known_manifest = PkgManifest(),
  ...)

\S4method{makeSeedMan}{parsedSessionInfo}(x,
  known_manifest = PkgManifest(), ...)

\S4method{makeSeedMan}{data.frame}(x, known_manifest = PkgManifest(),
  ...)
}
\arguments{
\item{x}{The object to generate a seeding manifest from, if missing, the
output from sessionInfo() is used.}

\item{known_manifest}{A manifest containing known locations of package sources.
makeSeedMan will attempt to determine locations of packages listed in x using both
known_manifest and official repositories.}

\item{...}{Currently unused.}
}
\description{
makeSeedMan
}
\examples{
man = makeSeedMan()
}
