% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accessors.R
\name{manifest}
\alias{manifest}
\alias{manifest,PkgManifest-method}
\title{manifest
Get or set the PkgManifest associated with an object.}
\usage{
manifest(x)
}
\description{
manifest
Get or set the PkgManifest associated with an object.
}

