% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Manifest}
\alias{Manifest}
\title{Manifest}
\usage{
Manifest(..., dep_repos = c(biocinstallRepos()))
}
\arguments{
\item{\dots{}}{Vectors containing package information. Passed to \code{\link{ManifestRow}}}

\item{dep_repos}{The dependency repos for the package.}
}
\description{
Create a package manifest
}

