% Generated by roxygen2 (4.0.2): do not edit by hand
\name{manifest}
\alias{manifest}
\alias{manifest,PkgManifest-method}
\title{manifest
Extract manifest data.frame associated with the manifest}
\usage{
manifest(x)
}
\description{
manifest
Extract manifest data.frame associated with the manifest
}

