% Generated by roxygen2 (4.0.1): do not edit by hand
\name{manifest_df}
\alias{manifest_df}
\alias{manifest_df<-}
\title{manifest_df}
\usage{
manifest_df(x, ...)

manifest_df(x) <- value
}
\arguments{
\item{x}{The object}

\item{\dots{}}{unused.}

\item{value}{A data.frame of package manifest information.
See \code{\link{ManifestRow}}}
}
\description{
Get or set the package location manifest (data.frame) associated with an
object
}

