% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{manifest_df}
\alias{manifest_df}
\alias{manifest_df,SessionManifest-method}
\alias{manifest_df,SessionManifest}
\alias{manifest_df,PkgManifest-method}
\alias{manifest_df,PkgManifest}
\alias{manifest_df<-}
\alias{manifest_df<-,SessionManifest-method}
\alias{manifest_df<-,SessionManifest}
\alias{manifest_df<-,PkgManifest-method}
\alias{manifest_df<-,PkgManifest}
\title{manifest_df}
\usage{
manifest_df(x, ...)

\S4method{manifest_df}{SessionManifest}(x, session_only = TRUE, ...)

\S4method{manifest_df}{PkgManifest}(x)

manifest_df(x) <- value

\S4method{manifest_df}{SessionManifest}(x) <- value

\S4method{manifest_df}{PkgManifest}(x) <- value
}
\arguments{
\item{x}{The object}

\item{...}{unused.}

\item{session_only}{Only return manifest rows associated with}

\item{value}{A data.frame of package manifest information.
See \code{\link{ManifestRow}}}
}
\description{
Get or set the package location manifest (data.frame) associated with an
object
}
