% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{manifest}
\alias{manifest}
\alias{manifest<-}
\alias{manifest,SessionManifest-method}
\alias{manifest,SessionManifest}
\alias{manifest<-,SessionManifest-method}
\alias{manifest<-,SessionManifest}
\title{Get or set the manifest associated with an object}
\usage{
manifest(x)

manifest(x) <- value

\S4method{manifest}{SessionManifest}(x)

\S4method{manifest}{SessionManifest}(x) <- value
}
\arguments{
\item{x}{An object which contains a manifest}

\item{value}{A PkgManifest}
}
\value{
A PkgManifest or SessionManifest object
}
\description{
Get or set manifest associated with an object
}
