% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manifest.R
\name{ManifestRow}
\alias{ManifestRow}
\title{ManifestRow}
\usage{
ManifestRow(name = NA_character_, url = NA_character_,
  type = NA_character_, branch = NA_character_, subdir = ".",
  extra = NA_character_)
}
\arguments{
\item{name}{name of the package}

\item{url}{location of the package sources}

\item{type}{type of location (svn, git, local, etc)}

\item{branch}{name of the branch to use to build the package}

\item{subdir}{subdirectory to use to build the package}

\item{extra}{currently ignored. extra commands for building or
installing the package}
}
\value{
A valid Package manifest data.frame
}
\description{
Create one or more rows of a manifest data.frame
}

