% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normalizePath2}
\alias{normalizePath2}
\title{normalizePath2}
\usage{
normalizePath2(path, follow.symlinks = FALSE)
}
\arguments{
\item{path}{The path to normalize}

\item{follow.symlinks}{Should symlinks (other than . and ..)
be resolved to their physical locations? (FALSE)}
}
\value{
The normalized path.
}
\description{
Attempt to normalize a relative path to an absolute one. Optionally without
resolving symlinks on non-Windows systems
}

