% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{notrack}
\alias{notrack}
\alias{notrack,NULL}
\alias{notrack,NULL-method}
\title{Notrack directory}
\usage{
notrack(repo)

\S4method{notrack}{`NULL`}(repo)
}
\arguments{
\item{repo}{The object.}
}
\value{
the path where retrieved package versions should be. If \code{repo}
is NULL, a notrack directory is constructed within a temp directory.
}
\description{
This function is not intended to be called directly by the user.
}

