% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{nrow}
\alias{nrow}
\alias{nrow,PkgManifest-method}
\alias{nrow,PkgManifest}
\alias{nrow,SessionManifest-method}
\alias{nrow,SessionManifest}
\title{Number of rows}
\usage{
nrow(x)

\S4method{nrow}{PkgManifest}(x)

\S4method{nrow}{SessionManifest}(x)
}
\arguments{
\item{x}{A tabular data structure.}
}
\value{
The number of rows in the structure
}
\description{
Number of rows
}
