% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{packages}
\alias{packages}
\alias{packages,SwitchrCtx-method}
\alias{packages,SwitchrCtx}
\title{packages}
\usage{
packages(seed)

\S4method{packages}{SwitchrCtx}(seed)
}
\arguments{
\item{seed}{A switchr context}
}
\description{
List the packages installed in a switchr context (library)
}
