% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionInfo.R
\name{parseSessionInfoString}
\alias{parseSessionInfoString}
\title{Parse text output from printing SessionInfo objects}
\usage{
parseSessionInfoString(string)
}
\arguments{
\item{string}{The text output from sessionInfo()}
}
\value{
a \code{parsedSessionInfo} object encoding the
information \code{string} (ie print output of a
\code{sessionInfo} call).
}
\description{
Parse text output from printing SessionInfo objects
}
