% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PkgSource-class}
\alias{PkgSource-class}
\alias{SVNSource-class}
\alias{GitSource-class}
\alias{GithubSource-class}
\alias{CVSSource-class}
\alias{LocalSource-class}
\alias{CRANSource-class}
\alias{BiocSource-class}
\alias{TarballSource-class}
\title{PkgSource}
\description{
An object representing the source location of a package. This is a virtual
used exlusively through its subclasses, which are used to differentiate the
different types of package source locations.
}
