% Generated by roxygen2 (4.0.1): do not edit by hand
\name{publishManifest}
\alias{publishManifest}
\title{publishManifest}
\usage{
publishManifest(manifest, dest = "./pkg_manifest.rman", ...)
}
\arguments{
\item{manifest}{The object to save as a serialized package or session
manifest. Defaulst to the currently in use switchr library.
A session manifest will be generated by libManifest as necessary.}

\item{dest}{The destination \code{manifest} will be published to. Typically
a character value indicating a file name (including path) to write to.}

\item{...}{Unused}
}
\value{
The name of the file written
}
\description{
Publish a package or session manifest to file.
}

