% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveManifest.R
\docType{methods}
\name{publishManifest}
\alias{publishManifest}
\alias{publishManifest,PkgManifest,character-method}
\alias{publishManifest,PkgManifest,character}
\alias{publishManifest,SessionManifest,character-method}
\alias{publishManifest,SessionManifest,character}
\alias{publishManifest,missing,ANY-method}
\alias{publishManifest,missing,ANY}
\alias{publishManifest,SwitchrCtx,ANY-method}
\alias{publishManifest,SwitchrCtx,ANY}
\title{publishManifest}
\usage{
publishManifest(manifest, dest = "./pkg_manifest.rman", ...)

\S4method{publishManifest}{PkgManifest,character}(manifest,
  dest = "./pkg_manifest.rman", ...)

\S4method{publishManifest}{SessionManifest,character}(manifest,
  dest = "./pkg_manifest.rman", ...)

\S4method{publishManifest}{missing,ANY}(manifest,
  dest = "./pkg_manifest.rman", ...)

\S4method{publishManifest}{SwitchrCtx,ANY}(manifest,
  dest = "./pkg_manifest.rman", ...)
}
\arguments{
\item{manifest}{The object to save as a serialized package or session
manifest. Defaulst to the currently in use switchr library.
A session manifest will be generated by libManifest as necessary.}

\item{dest}{The destination \code{manifest} will be published to. Typically
a character value indicating a file name (including path) to write to.}

\item{\dots}{Unused}
}
\value{
The name of the file written
}
\description{
Publish a package or session manifest to file.
}
