% Generated by roxygen2 (4.0.2): do not edit by hand
\name{removeLib}
\alias{removeLib}
\title{Remove a switchr library and update the manifest of existing libraries}
\usage{
removeLib(name = NULL, repos = NULL, compEnv = NULL, fromStack = FALSE)
}
\description{
Remove a switchr library and update the manifest of existing libraries
}

