% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{RepoSubset}
\alias{RepoSubset}
\alias{RepoSubset-class}
\title{RepoSubset}
\usage{
RepoSubset(repos, pkgs, default_name)
}
\arguments{
\item{repos}{The traditional repositories to select the packages from}

\item{pkgs}{The packages included in the subset}

\item{default_name}{The default name to use when the RepoSubset is used to
seed a switchr context}
}
\description{
An object that represents a subset of packages available in a repo. When switched to, switchr will default to only installing the
specified packages, rather than all packages in the repository.
}
