% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandbtmp.R
\name{rVersionManifest}
\alias{rVersionManifest}
\title{rVersionManifest}
\usage{
rVersionManifest(vers, curr_avail = available.packages())
}
\arguments{
\item{vers}{The version of R to create a manifest for}

\item{curr_avail}{The output from available.packages(). Used to identify
whether the necessary version is in the CRAN archive or normal repository}
}
\value{
A SessionManifest object
}
\description{
Create a Pkg manifest which points to tarballs representing the
cohort of packages associated with a particular release of R
}
\examples{
\dontrun{
man = rVersionManifest("3.1.1")
man
}
}
\references{
"Gabor Csardi" (2014). crandb: Query the unofficial CRAN metadata
 database. R package version 1.0.0. https://github.com/metacran/crandb

Becker G, Barr C, Gentleman R, Lawrence M; Enhancing Reproducibility and Collaboration via Management of R Package Cohorts. Journal of Statistical Software, 81(1). 2017. doi: 10.18637/jss.v082.i01
}
\author{
Gabriel Becker
}
