% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{sh_init_script}
\alias{sh_init_script}
\alias{sh_init_script,SwitchrParam-method}
\alias{sh_init_script,SwitchrParam}
\alias{sh_init_script<-}
\alias{sh_init_script<-,SwitchrParam-method}
\alias{sh_init_script<-,SwitchrParam,ANY}
\title{shell init}
\usage{
sh_init_script(x)

\S4method{sh_init_script}{SwitchrParam}(x)

sh_init_script(x) <- value

\S4method{sh_init_script}{SwitchrParam}(x) <- value
}
\arguments{
\item{x}{An object associated with a SwitchrParam object}

\item{value}{The new value.}
}
\description{
Set or Retrieve the shell initialization script for an object
}
