% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flushSession.R
\docType{data}
\name{switchDeps}
\alias{switchDeps}
\title{switchrDeps}
\format{
An object of class \code{character} of length 20.
}
\usage{
switchDeps
}
\value{
the base packages, plus the set of packages switchr itself has dependencies on
}
\description{
The base packages, as well as switchr and its dependencies.
}
\keyword{datasets}
