% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SwitchrParam.R
\name{SwitchrParam}
\alias{SwitchrParam}
\alias{SwitchrParam-class}
\title{SwitchrParam
A contstructor for a SwitchrParam object representing a number of common
parameters understood by the switchr framework}
\usage{
SwitchrParam(logfun = function(...) NULL, shell_init = character(),
  archive_timing = 2, archive_retries = 2)
}
\arguments{
\item{logfun}{The function to be called to write to logs}

\item{shell_init}{A character containing the location of a shell script to
be sourced before any system commands.}
}
\value{
A SwitchrParam object.
}
\description{
SwitchrParam
A contstructor for a SwitchrParam object representing a number of common
parameters understood by the switchr framework
}
\author{
Gabriel Becker
}

