% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/SwitchrParam.R
\docType{class}
\name{SwitchrParam-class}
\alias{SwitchrParam-class}
\alias{SwitchrParam}
\title{SwitchrParam}
\usage{
SwitchrParam(logfun = function(...) NULL, shell_init = character(),
  archive_timing = 2, archive_retries = 2, dl_method,
  shell_timing = 1)
}
\arguments{
\item{logfun}{The function to be called to write to logs}

\item{shell_init}{A character containing the location of a shell script to
be sourced before any system commands.}

\item{archive_timing}{The timeout after downloading a package from the CRAN Archive.}

\item{archive_retries}{Number of times to retry retrieving a package from the CRAN Archive.}

\item{dl_method}{The download method to use when retrieve package
source files. See \code{\link{download.file}} If none is specified,
the method defaults to "curl" if the RCurl package is installed and "auto" otherwise.}

\item{shell_timing}{numeric. The numer of seconds to wait between certain shell commands.
Defaults to 1, this should only need to be changed in the case of, e.g., networked drive latency issues.}
}
\value{
A SwitchrParam object.
}
\description{
A contstructor for a SwitchrParam object representing a number of common
parameters understood by the switchr framework
}
\author{
Gabriel Becker
}
