% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{system_w_init}
\alias{system_w_init}
\title{system_w_init}
\usage{
system_w_init(cmd, dir, init = character(), args = NULL, env = NULL,
  ..., param = SwitchrParam())
}
\arguments{
\item{cmd}{The text of the command. Must be length 1.}

\item{dir}{The directory that the command should be executed in. The working directory will be temporarily changed to this dir,
but will be changed back upon exit of system_w_init.}

\item{init}{(optional) a character value indicating the
location of an initialization shell script.}

\item{args}{character. Arguments to be passed to the command}

\item{env}{character. Environmental variables to be set when running the command}

\item{\dots}{additional parameters passed directly to \code{\link{system}}.}

\item{param}{A SwitchrParam object. The shell initialization
script associated with this object is used when \code{init} is
not specified (length 0).}
}
\value{
Depends, see \code{\link{system}} for details.
}
\description{
Run a system command with an optional intialization script (e.g. a .bashrc
sourced first).
}
