% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCtx.R
\name{updateManifest}
\alias{updateManifest}
\title{updateManifest}
\usage{
updateManifest()
}
\value{
NULL, used for it's side-effect of updating the switchr library
metadata cache.
}
\description{
Update the cached information regarding available switchr libraries.
}
