% Generated by roxygen2 (4.0.1): do not edit by hand
\name{versions_df}
\alias{versions_df}
\alias{versions_df<-}
\title{versions_df}
\usage{
versions_df(x)

versions_df(x) <- value
}
\arguments{
\item{x}{An object containing package version information}

\item{value}{A data.frame of package version information.}
}
\description{
Get or set the the versions information in a SessionManifest
}

