% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{versions_df}
\alias{versions_df}
\alias{versions_df,SessionManifest-method}
\alias{versions_df,SessionManifest}
\alias{versions_df<-}
\alias{versions_df<-,SessionManifest-method}
\alias{versions_df<-,SessionManifest}
\title{versions_df}
\usage{
versions_df(x)

\S4method{versions_df}{SessionManifest}(x)

versions_df(x) <- value

\S4method{versions_df}{SessionManifest}(x) <- value
}
\arguments{
\item{x}{An object containing package version information}

\item{value}{A data.frame of package version information.}
}
\description{
Get or set the the versions information in a SessionManifest
}
